+++ src/test/groovy/ru/vyarus/java/generics/resolver/cases/methodgeneric/support/SubMethodGenericCase.groovy
@@ -9,4 +9,8 @@ class SubMethodGenericCase<O> {
     public <T extends O> T testSub(Class<T> arg, T arg2) {
         return null;
     }
+
+    public <T extends O, K extends T> T testSub2(Class<T> arg, K arg2) {
+        return null;
+    }
 }